refactor(webpack): Update build loaders
