+++ package.json
@@ -24,9 +24,10 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@babel/core": "^7.23.2",
-    "@babel/preset-react": "^7.22.15",
-    "@babel/preset-typescript": "^7.23.2",
+    "@babel/core": "^7.28.3",
+    "@babel/preset-env": "^7.28.3",
+    "@babel/preset-react": "^7.27.1",
+    "@babel/preset-typescript": "^7.27.1",
     "@commitlint/cli": "^19.8.1",
     "@commitlint/config-conventional": "^19.8.1",
     "@cypress/code-coverage": "^3.14.5",
@@ -44,7 +45,9 @@
     "@types/webpack": "^5.28.5",
     "@typescript-eslint/eslint-plugin": "^8.38.0",
     "@typescript-eslint/parser": "^8.38.0",
+    "babel-loader": "^10.0.0",
     "babel-plugin-istanbul": "^7.0.0",
+    "babel-plugin-jsx-remove-data-test-id": "^3.0.0",
     "cypress": "^14.5.4",
     "dotenv": "^17.2.1",
     "eslint": "^9.32.0",
