+++ src/constants/index.ts
@@ -1,20 +1,13 @@
 export * from "./cacheConstants";
 export * from "./calendarEvents";
-export * from "./calendarEventTest";
 export * from "./conditions";
-export * from "./displayOfElementsTestConstants";
 export * from "./envConstants";
 export * from "./errors";
-export * from "./geolocationIsTurnOnTestConstants";
 export * from "./includesConstants";
 export * from "./localstorageConstants";
-export * from "./notFoundPageTestConstants";
 export * from "./numericalConstants";
-export * from "./permissionDeniedTestConstants";
 export * from "./routes";
-export * from "./searchWithSuccessConstant";
 export * from "./swiper";
-export * from "./switchTypeOfWeatherConstants";
 export * from "./theme";
 export * from "./time";
 export * from "./UI";
