--- src/mocks/emptyArrayMock.ts
@@ -1,1 +1,0 @@
-export const EMPTY_ARRAY_MOCK: unknown = [];
