+++ src/mocks/events.ts
@@ -1,3 +1,13 @@
+export const EVENTS_MOCK = [
+  {
+    id: "event1",
+    summary: "JS internships Daily Meeting",
+    start: { dateTime: "2025-08-20T10:00:00+03:00" },
+    end: { dateTime: "2025-08-20T10:30:00+03:00" },
+    description: "Test meeting description",
+  },
+];
+
 export const CALENDAR_EVENTS_TEST = {
   DESCRIPTION: "Calendar Events Test",
   IT: {
