--- src/mocks/eventsMock.ts
@@ -1,9 +1,0 @@
-export const EVENTS_MOCK = [
-  {
-    id: "event1",
-    summary: "JS internships Daily Meeting",
-    start: { dateTime: "2025-08-20T10:00:00+03:00" },
-    end: { dateTime: "2025-08-20T10:30:00+03:00" },
-    description: "Test meeting description",
-  },
-];
