+++ src/mocks/index.ts
@@ -1,5 +1,8 @@
-export * from "./emptyArrayMock";
-export * from "./eventsMock";
-export * from "./geolocationIsTurnOnMock";
-export * from "./searchWithSuccessResponseMock";
-export * from "./switchTypeOfWeatherMock";
+export * from "./displayOfElements";
+export * from "./events";
+export * from "./geolocationIsTurnOn";
+export * from "./notFoundPage";
+export * from "./permissionDenied";
+export * from "./searchWithNoResult";
+export * from "./searchWithSuccessResponse";
+export * from "./switchTypeOfWeather";
